post list

+++ app/src/main/java/com/jgc/myjsonplaceholder/data/repositories/ListRepositoryImpl.kt
@@ -22,7 +22,7 @@ class ListRepositoryImpl : ListRepository {
 
 fun List<PostItemDto>.toDomainLayer(): List<Post> = map {
     Post(
-        id = it.id,
+        id = it.id.toLong(),
         body = it.body,
         title = it.title,
         userId = it.userId
